订单详情添加之前commit

+++ renren-api/src/main/java/io/renren/rabbitmq/MQReceiver.java
@@ -41,7 +41,7 @@ public class MQReceiver {
         long goodsId = mm.getGoodsId();
 
         GoodsVo goods = goodsService.getGoodsVoByGoodsId(goodsId);
-        int stock = goods.getStockCount();
+        long stock = goods.getStockCount();
         if (stock <= 0) {
             return;
         }
@@ -56,25 +56,5 @@ public class MQReceiver {
         miaoshaOrderService.secKill(goods, user);
     }
 
-//		@RabbitListener(queues=MQConfig.QUEUE)
-//		public void receive(String message) {
-//			log.info("receive message:"+message);
-//		}
-//
-//		@RabbitListener(queues=MQConfig.TOPIC_QUEUE1)
-//		public void receiveTopic1(String message) {
-//			log.info(" topic  queue1 message:"+message);
-//		}
-//
-//		@RabbitListener(queues=MQConfig.TOPIC_QUEUE2)
-//		public void receiveTopic2(String message) {
-//			log.info(" topic  queue2 message:"+message);
-//		}
-//
-//		@RabbitListener(queues=MQConfig.HEADER_QUEUE)
-//		public void receiveHeaderQueue(byte[] message) {
-//			log.info(" header  queue message:"+new String(message));
-//		}
-//
 
 }
